requirements.txt
cachelib==0.10.2
certifi @ file:///C:/ci_311/certifi_1676421629969/work/certifi
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
flask-http-middleware==0.2.1
Flask-Session==0.4.0
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
pycparser==2.21
pyOpenSSL==23.1.1
python-dotenv==1.0.0
Werkzeug==2.2.3
wincertstore==0.2
WTForms==3.0.1
